Move module declaration to a separate TestModules interface
